pyproject.toml
[project]
name = "kiri"
version = "0.1.0"
description = "Kiri — voice-to-text assistant powered by OpenVINO Whisper"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "huggingface-hub[cli]>=0.36.2",
    "nncf>=2.19.0",
    "numpy",
    "openvino>=2025.4.1",
    "openvino-genai>=2025.4.1.0",
    "optimum-intel>=1.27.0",
    "pygobject>=3.54.5",
    "scipy>=1.17.1",
    "sounddevice>=0.5.5",
    "transformers>=4.57.6",
]

[project.scripts]
kiri = "kiri.cli.main:main"
kiri-popup = "kiri.cli.gui_main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/kiri"]
